+++ sigla-ejb/src/main/java/it/cnr/contab/config00/bulk/Configurazione_cnrBulk.java
@@ -176,6 +176,12 @@ public class Configurazione_cnrBulk extends Configurazione_cnrBase {
     public final static String PK_STIPENDI = "STIPENDI";
     public final static String SK_CODICE_BOLLO = "CODICE_BOLLO";
 
+    public final static String PK_IMPEGNI = "IMPEGNI";
+    public final static String SK_IMPEGNI_PLURIENNALI = "IMPEGNI_PLURIENNALI";
+
+    public final static String PK_ACCERTAMENTI = "ACCERTAMENTI";
+    public final static String SK_ACCERTAMENTI_PLURIENNALI = "ACCERTAMENTI_PLURIENNALI";
+
     public Configurazione_cnrBulk() {
         super();
     }