+++ sigla-ejb/src/main/java/it/cnr/contab/doccont00/core/bulk/AccertamentoBulk.java
@@ -110,6 +110,10 @@ public class AccertamentoBulk extends AccertamentoBase implements IDocumentoCont
 	private BulkList<AllegatoGenericoBulk> archivioAllegati = new BulkList<AllegatoGenericoBulk>();
 	private boolean enableVoceNext = false;
 
+	private BulkList<Accertamento_pluriennaleBulk> accertamentiPluriennali = new BulkList<Accertamento_pluriennaleBulk>();
+
+
+
 public AccertamentoBulk() {
 	super();
 }
@@ -209,7 +213,8 @@ public BulkCollection[] getBulkLists() {
 	// Metti solo le liste di oggetti che devono essere resi persistenti
 	
 	 return new it.cnr.jada.bulk.BulkCollection[] { 
-			accertamento_scadenzarioColl, archivioAllegati, pdgVincoliColl, accertamentoVincoliPerentiColl };
+			accertamento_scadenzarioColl, archivioAllegati, pdgVincoliColl,
+			 accertamentoVincoliPerentiColl, accertamentiPluriennali };
 }
 /**
  * Insert the method's description here.
@@ -1368,6 +1373,16 @@ public void setCd_cds(java.lang.String cd_cds) {
 	public BigDecimal getImportoNonIncassato() {
 		return this.getAccertamento_scadenzarioColl().stream().map(e->e.getImportoNonIncassato()).reduce((x, y)->x.add(y)).orElse(BigDecimal.ZERO);
 	}
+	private BulkList<Accertamento_pluriennaleBulk> clonaAccertamentiPluriennali(AccertamentoBulk accertamento,it.cnr.jada.action.ActionContext context){
+		if ( this.getAccertamentiPluriennali()==null || this.getAccertamentiPluriennali().isEmpty())
+			return this.getAccertamentiPluriennali();
+		BulkList<Accertamento_pluriennaleBulk> pluriennali= new BulkList<Accertamento_pluriennaleBulk>();
+		for ( Accertamento_pluriennaleBulk p:this.getAccertamentiPluriennali()){
+			Accertamento_pluriennaleBulk n = p.clone(accertamento,context);
+			pluriennali.add( n);
+		}
+		return pluriennali;
+	}
 	public Object clona(it.cnr.jada.util.action.CRUDBP bp,it.cnr.jada.action.ActionContext context) {
 		AccertamentoBulk nuovo = null;
 		try {
@@ -1405,6 +1420,27 @@ public void setCd_cds(java.lang.String cd_cds) {
 		nuovo.setCd_uo_origine( getCd_uo_origine());
 		nuovo.setCd_cds_origine( getCd_cds_origine());
 		nuovo.setFl_netto_sospeso( getFl_netto_sospeso());
+		// da verificare se vanno clonati
+		//nuovo.setAccertamentiPluriennali(clonaAccertamentiPluriennali(nuovo,context));
 		return nuovo;
 	}
+
+	public BulkList<Accertamento_pluriennaleBulk> getAccertamentiPluriennali() {
+		return accertamentiPluriennali;
+	}
+	public void setAccertamentiPluriennali(BulkList<Accertamento_pluriennaleBulk> accertamentiPluriennali) {
+		this.accertamentiPluriennali = accertamentiPluriennali;
+	}
+
+	public int addToAccertamentiPluriennali(Accertamento_pluriennaleBulk dett){
+		dett.setAccertamento(this);
+		getAccertamentiPluriennali().add(dett);
+		return getAccertamentiPluriennali().size()-1;
+	}
+
+	public Accertamento_pluriennaleBulk removeFromAccertamentiPluriennali(int index) {
+		Accertamento_pluriennaleBulk dett = (Accertamento_pluriennaleBulk)getAccertamentiPluriennali().remove(index);
+		return dett;
+	}
+
 }