+++ sigla-ejb/src/main/java/it/cnr/contab/progettiric00/core/bulk/ProgettoHome.java
@@ -29,7 +29,6 @@ import it.cnr.contab.anagraf00.core.bulk.TerzoBulk;
 import it.cnr.contab.config00.bulk.*;
 import it.cnr.contab.config00.contratto.bulk.ContrattoBulk;
 import it.cnr.contab.config00.contratto.bulk.ContrattoHome;
-import it.cnr.contab.config00.contratto.bulk.ContrattoBulk;
 import it.cnr.contab.config00.latt.bulk.WorkpackageBulk;
 import it.cnr.contab.config00.pdcfin.bulk.Elemento_voceHome;
 import it.cnr.contab.config00.sto.bulk.*;
@@ -1038,6 +1037,15 @@ public class ProgettoHome extends BulkHome {
 		ContrattoHome dettHome = (ContrattoHome)getHomeCache().getHome(ContrattoBulk.class);
 		SQLBuilder sql = dettHome.createSQLBuilder();
 		sql.addClause(FindClause.AND,"pg_progetto",SQLBuilder.EQUALS,pgProgetto);
+		return dettHome.fetchAll(sql);
+	}
+
+	public java.util.Collection findAnagraficheProgetto(it.cnr.jada.UserContext userContext, ProgettoBulk progetto) throws PersistencyException {
+		PersistentHome dettHome = getHomeCache().getHome(Progetto_anagraficoBulk.class);
+		SQLBuilder sql = dettHome.createSQLBuilder();
+		sql.addSQLClause(FindClause.AND, "PG_PROGETTO", SQLBuilder.EQUALS, progetto.getPg_progetto());
+
+		sql.setOrderBy("ID_PRG_ANAGRAFICO", it.cnr.jada.util.OrderConstants.ORDER_ASC);
 		return dettHome.fetchAll(sql);
 	}